Add more heartbeat tests
